requirements.txt
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
contourpy==1.3.2
cycler==0.12.1
fonttools==4.57.0
google-auth==2.39.0
google-auth-oauthlib==1.2.1
gspread==6.2.0
idna==3.10
kiwisolver==1.4.8
matplotlib==3.10.1
numpy==2.2.4
oauthlib==3.2.2
packaging==24.2
pillow==11.2.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.17.0
urllib3==2.4.0
